+++ .github/workflows/main.yml
@@ -27,7 +27,6 @@ jobs:
       - name: Build and test
         run: |
           sudo apt-get update 
-          sudo apt-get install -y googletest
           sudo apt-get install -y apt-utils \
           build-essential \
           cmake \
