+++ cases_high_level_simulation/test_unit/input/ball_mass.stl
