+++ cases_high_level_simulation/test_unit/input/cylinder.stl
